// legacy_web_angular/src/app/interfaces/i-payment-detail.ts
export interface IIncomingPaymentDetail {
  DocNumOinv:number,
  DocNumPay:number,
  CashSum: number;
  CashSumFC: number;
  TrsfrSum: number;
  TrsfrSumFC: number;
  CreditCards: ICreditVouchersDetail[];
}
export interface ICreditVouchersDetail {
  Id: number;
  CreditCard: number;
  VoucherNum: string;
  CreditSum: number;
  CardValid:Date;
  CollectionDate: Date;
  Account: string;
  IsManualEntry:boolean;
}

export interface IDataForDetailPay{
  DocNumPay:number;
  DocNumOinv:number;
  DocEntryPay:number;
}